oops forgot these 2 places in renaming
