+++ server/neptune/workflows/internal/terraform/job/runner.go
@@ -217,7 +217,7 @@ func (r *JobRunner) apply(executionCtx *ExecutionContext, planFile string, step 
 
 func (r *JobRunner) plan(ctx *ExecutionContext, mode *terraform.PlanMode, workflowMode terraform.WorkflowMode, extraArgs []string) (activities.TerraformPlanResponse, error) {
 	if workflowMode == terraform.Adhoc {
-		// Admin mode doesn't need to run a plan.
+		// Adhoc mode doesn't need to run a plan.
 		return activities.TerraformPlanResponse{}, nil
 	}
 	var resp activities.TerraformPlanResponse
